+++ packages/firebase-dart/lib/src/interop/auth_interop.dart
@@ -10,25 +10,30 @@ import 'firebase_interop.dart';
 @JS('Auth')
 abstract class AuthJsImpl {
   external AppJsImpl get app;
-  external void set app(AppJsImpl a);
-  external UserJsImpl get currentUser;
-  external void set currentUser(UserJsImpl u);
   external PromiseJsImpl applyActionCode(String code);
   external PromiseJsImpl<ActionCodeInfo> checkActionCode(String code);
   external PromiseJsImpl confirmPasswordReset(String code, String newPassword);
   external PromiseJsImpl<UserJsImpl> createUserWithEmailAndPassword(
       String email, String password);
+  external UserJsImpl get currentUser;
   external PromiseJsImpl<List<String>> fetchProvidersForEmail(String email);
   external PromiseJsImpl<UserCredentialJsImpl> getRedirectResult();
   external Func0 onAuthStateChanged(nextOrObserver,
       [Func1 opt_error, Func0 opt_completed]);
+  external Func0 onIdTokenChanged(nextOrObserver,
+      [Func1 opt_error, Func0 opt_completed]);
   external PromiseJsImpl sendPasswordResetEmail(String email);
+  external PromiseJsImpl<UserCredentialJsImpl>
+      signInAndRetrieveDataWithCredential(AuthCredential credential);
   external PromiseJsImpl<UserJsImpl> signInAnonymously();
   external PromiseJsImpl<UserJsImpl> signInWithCredential(
       AuthCredential credential);
   external PromiseJsImpl<UserJsImpl> signInWithCustomToken(String token);
   external PromiseJsImpl<UserJsImpl> signInWithEmailAndPassword(
       String email, String password);
+  @Deprecated('not impld')
+  external PromiseJsImpl signInWithPhoneNumber(
+      String phoneNumber, /* ApplicationVerifier */ applicationVerifier);
   external PromiseJsImpl<UserCredentialJsImpl> signInWithPopup(
       AuthProviderJsImpl provider);
   external PromiseJsImpl signInWithRedirect(AuthProviderJsImpl provider);
@@ -43,14 +48,8 @@ abstract class AuthJsImpl {
 /// See: <https://firebase.google.com/docs/reference/js/firebase.auth.AuthCredential>.
 @JS()
 abstract class AuthCredential {
-  /// _Deprecated: Use [providerId] instead._
-  @deprecated
-  external String get provider;
-
   /// The authentication provider ID for the credential.
   external String get providerId;
-  external String get accessToken;
-  external String get secret;
 }
 
 @JS('AuthProvider')
@@ -102,6 +101,19 @@ class TwitterAuthProviderJsImpl extends AuthProviderJsImpl {
   external static AuthCredential credential(String token, String secret);
 }
 
+@JS('PhoneAuthProvider')
+class PhoneAuthProviderJsImpl extends AuthProviderJsImpl {
+  external factory PhoneAuthProviderJsImpl();
+  external static String get PROVIDER_ID;
+
+  /// The user's [phoneNumber] in E.164 format (e.g. +16505550101).
+  external verifyPhoneNumber(String phoneNumber, applicationVerifier);
+
+  // https://firebase.google.com/docs/reference/js/firebase.auth.PhoneAuthProvider#.credential
+  external static PromiseJsImpl<AuthCredential> credential(
+      String verificationId, String verificationCode);
+}
+
 /// A response from [Auth.checkActionCode].
 ///
 /// See: <https://firebase.google.com/docs/reference/js/firebase.auth.ActionCodeInfo>.
@@ -127,16 +139,21 @@ class ActionCodeEmail {
   external String get email;
 }
 
+/// https://firebase.google.com/docs/reference/js/firebase.auth#.UserCredential
 @JS()
 @anonymous
 class UserCredentialJsImpl {
+  external AdditionalUserInfoJsImpl get additionalUserInfo;
   external UserJsImpl get user;
-  external void set user(UserJsImpl u);
   external AuthCredential get credential;
-  external void set credential(AuthCredential c);
   external String get operationType;
-  external void set operationType(String t);
+}
 
-  external factory UserCredentialJsImpl(
-      {UserJsImpl user, AuthCredential credential, String operationType});
+/// https://firebase.google.com/docs/reference/js/firebase.auth#.AdditionalUserInfo
+@JS()
+@anonymous
+class AdditionalUserInfoJsImpl {
+  external String get providerId;
+  external Object get profile;
+  external String get username;
 }